requirements.txt
Django==1.6.5
MySQL-python==1.2.5
PyYAML==3.11
South==1.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-tastypie==0.11.1
gunicorn==19.0.0
inflect==0.2.4
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
static3==0.5.1
wsgiref==0.1.2
